Added SwipeItem and fixed props
